test(token): refactor test helpers and improve coverage
- Move writeECKeyFile helper to shared test_helpers.go for reuse across test files
- Add TestParseInvalidToken to cover error handling in Parse function
- Improves test coverage for ECDSA token implementation
